livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-ai-coustics~=0.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.5.4",
    "python-dotenv>=1.2.2",
]
